requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
diffprivlib==0.6.6
filelock==3.20.3
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
Flask-Logging==0.1.3
Flask-Mail==0.10.0
Flask-Uploads==0.2.1
fsspec==2026.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
mpmath==1.3.0
mysql-connector-python==9.5.0
networkx==3.6.1
numpy==2.4.1
pandas==2.3.3
pycparser==2.23
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
Werkzeug==3.1.5
wheel==0.45.1
